+++ src/Symfony/Component/Security/Core/Validator/Constraint/UserPassword.php
@@ -11,18 +11,19 @@
 
 namespace Symfony\Component\Security\Core\Validator\Constraint;
 
-use Symfony\Component\Validator\Constraint;
+use Symfony\Component\Security\Core\Validator\Constraints\UserPassword as BaseUserPassword;
 
 /**
  * @Annotation
+ *
+ * @deprecated Deprecated since version 2.2, to be removed in 2.3.
  */
-class UserPassword extends Constraint
+class UserPassword extends BaseUserPassword
 {
-    public $message = 'This value should be the user current password.';
-    public $service = 'security.validator.user_password';
-
-    public function validatedBy()
+    public function __construct($options = null)
     {
-        return $this->service;
+        trigger_error('UserPassword class in Symfony\Component\Security\Core\Validator\Constraint namespace is deprecated since version 2.2 and will be removed in 2.3. Use the Symfony\Component\Security\Core\Validator\Constraints\UserPassword class instead.', E_USER_DEPRECATED);
+
+        parent::__construct($options);
     }
 }
